Move altapi router file
